requirements.txt
tensorflow-cpu
numpy
Flask
uwsgi
Pillow
